Move SyncSettingsFragment to a new file
